# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Friday Food Club API — drops, orders, chefs, admin, subscriptions
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /fulfillment/verify:
    post:
      operationId: verifyPickup
      tags: [fulfillment]
      summary: Chef scans buyer QR token — releases escrow split
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PickupVerification' }
      responses:
        '200':
          description: Escrow released
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FulfillmentResult' }
...
  /fulfillment/wallet/{chefId}:
    get:
      operationId: getChefWallet
      tags: [fulfillment]
      summary: Chef wallet balance and payout history
      parameters:
        - name: chefId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Wallet
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ChefWallet' }
components:
  schemas:
...
    PickupVerification:
      type: object
      required: [pickupToken, chefId]
      properties:
        pickupToken: { type: string }
        chefId:      { type: string }
    FulfillmentResult:
      type: object
      properties:
        orderId:           { type: string }
        dropId:            { type: string }
        dropTitle:         { type: string }
        paymentMethod:     { type: string, enum: [DIGITAL, CASH] }
        gross:             { type: number }
        chefShare:         { type: number }
        platformShare:     { type: number }
        chefWalletBalance: { type: number }
        cashCollected:     { type: ["number", "null"], description: "For CASH orders: gross amount collected by chef in cash" }
        status:            { type: string }
...
    WalletTransaction:
      type: object
      properties:
        orderId:       { type: string }
        dropTitle:     { type: string }
        paymentMethod: { type: string, enum: [DIGITAL, CASH] }
        chefShare:     { type: number }
        platformShare: { type: number }
        cashCollected: { type: ["number", "null"] }
        walletEffect:  { type: number, description: "Positive = credit (digital payout); Negative = debit (cash fee)" }
        fulfilledAt:   { type: ["string", "null"] }
...
    ChefWallet:
      type: object
      properties:
        walletBalance:      { type: number }
        freezeThreshold:    { type: number, description: "Wallet balance below which drop creation is blocked" }
        isFrozen:           { type: boolean, description: "True when walletBalance < freezeThreshold" }
        cashDebt:           { type: number, description: "Absolute value of negative balance (0 if balance is positive)" }
        totalEarnings:      { type: number, description: "All-time sum of net chefShare for all fulfilled orders (never truncated)" }
        recentPayouts:
          type: array
          description: "Alias for recentTransactions — backwards compatible"
          items:
            $ref: "#/components/schemas/WalletTransaction"
        recentTransactions:
          type: array
          items:
            $ref: "#/components/schemas/WalletTransaction"